Refine canvas store and provider settings
